delete presentation - ui directory
