+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.App">
-        <activity android:name="com.nexters.lettero.presentation.ui.main.activity.MainActivity">
+        <activity android:name="com.nexters.lettero.presentation.main.activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
