+++ app/src/main/java/com/nexters/lettero/presentation/base/ViewModel.kt
@@ -1,0 +1,3 @@
+package com.nexters.lettero.presentation.base
+
+interface ViewModel {}
