+++ app/src/main/java/com/nexters/lettero/presentation/intro/viewmodel/IntroViewModel.kt
@@ -1,0 +1,5 @@
+package com.nexters.lettero.presentation.intro.viewmodel
+
+import com.nexters.lettero.presentation.base.ViewModel
+
+class IntroViewModel : ViewModel {}
