+++ app/src/main/java/com/nexters/lettero/presentation/login/viewmodel/LoginViewModel.kt
@@ -1,0 +1,5 @@
+package com.nexters.lettero.presentation.login.viewmodel
+
+import com.nexters.lettero.presentation.base.ViewModel
+
+class LoginViewModel : ViewModel {}
