+++ app/src/main/java/com/nexters/lettero/presentation/main/viewmodel/InBoxViewModel.kt
@@ -1,0 +1,6 @@
+package com.nexters.lettero.presentation.main.viewmodel
+
+import com.nexters.lettero.presentation.base.ViewModel
+
+class InBoxViewModel : ViewModel {
+}
