+++ app/src/main/java/com/nexters/lettero/presentation/main/viewmodel/MainViewModel.kt
@@ -1,0 +1,5 @@
+package com.nexters.lettero.presentation.main.viewmodel
+
+import com.nexters.lettero.presentation.base.ViewModel
+
+class MainViewModel : ViewModel {}
