+++ app/src/main/java/com/nexters/lettero/presentation/main/viewmodel/OutBoxViewModel.kt
@@ -1,0 +1,6 @@
+package com.nexters.lettero.presentation.main.viewmodel
+
+import com.nexters.lettero.presentation.base.ViewModel
+
+class OutBoxViewModel : ViewModel {
+}
