+++ app/src/main/java/com/nexters/lettero/presentation/message/viewmodel/SendMessageViewModel.kt
@@ -1,0 +1,5 @@
+package com.nexters.lettero.presentation.message.viewmodel
+
+import com.nexters.lettero.presentation.base.ViewModel
+
+class SendMessageViewModel : ViewModel {}
