+++ app/src/main/java/com/nexters/lettero/presentation/mypage/viewmodel/MyPageViewModel.kt
@@ -1,0 +1,5 @@
+package com.nexters.lettero.presentation.mypage.viewmodel
+
+import com.nexters.lettero.presentation.base.ViewModel
+
+class MyPageViewModel : ViewModel {}
