--- app/src/main/java/com/nexters/lettero/presentation/ui/base/BaseFragment.kt
@@ -1,31 +1,0 @@
-package com.nexters.lettero.presentation.ui.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import androidx.fragment.app.Fragment
-import com.nexters.lettero.BR
-
-abstract class BaseFragment<VB : ViewDataBinding, VM : ViewModel> : Fragment() {
-    protected lateinit var binding: VB
-    protected abstract var viewModel: ViewModel
-
-    @get:LayoutRes
-    protected abstract val layoutRes: Int
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
-        binding.setVariable(BR.viewModel, viewModel)
-        binding.executePendingBindings()
-        binding.lifecycleOwner = this.viewLifecycleOwner
-        return binding.root
-    }
-}
