--- app/src/main/java/com/nexters/lettero/presentation/ui/base/ViewModel.kt
@@ -1,3 +1,0 @@
-package com.nexters.lettero.presentation.ui.base
-
-interface ViewModel {}
