--- app/src/main/java/com/nexters/lettero/presentation/ui/intro/viewmodel/IntroViewModel.kt
@@ -1,5 +1,0 @@
-package com.nexters.lettero.presentation.ui.intro.viewmodel
-
-import com.nexters.lettero.presentation.ui.base.ViewModel
-
-class IntroViewModel : ViewModel {}
