--- app/src/main/java/com/nexters/lettero/presentation/ui/login/viewmodel/LoginViewModel.kt
@@ -1,5 +1,0 @@
-package com.nexters.lettero.presentation.ui.login.viewmodel
-
-import com.nexters.lettero.presentation.ui.base.ViewModel
-
-class LoginViewModel : ViewModel {}
