--- app/src/main/java/com/nexters/lettero/presentation/ui/main/fragment/OutBoxFragment.kt
@@ -1,12 +1,0 @@
-package com.nexters.lettero.presentation.ui.main.fragment
-
-import com.nexters.lettero.R
-import com.nexters.lettero.databinding.FragmentOutBoxBinding
-import com.nexters.lettero.presentation.ui.base.BaseFragment
-import com.nexters.lettero.presentation.ui.base.ViewModel
-import com.nexters.lettero.presentation.ui.main.viewmodel.OutBoxViewModel
-
-class OutBoxFragment : BaseFragment<FragmentOutBoxBinding, OutBoxViewModel>() {
-    override val layoutRes: Int = R.layout.fragment_out_box
-    override var viewModel: ViewModel = OutBoxViewModel()
-}
