--- app/src/main/java/com/nexters/lettero/presentation/ui/main/viewmodel/MainViewModel.kt
@@ -1,5 +1,0 @@
-package com.nexters.lettero.presentation.ui.main.viewmodel
-
-import com.nexters.lettero.presentation.ui.base.ViewModel
-
-class MainViewModel : ViewModel {}
