--- app/src/main/java/com/nexters/lettero/presentation/ui/main/viewmodel/OutBoxViewModel.kt
@@ -1,6 +1,0 @@
-package com.nexters.lettero.presentation.ui.main.viewmodel
-
-import com.nexters.lettero.presentation.ui.base.ViewModel
-
-class OutBoxViewModel : ViewModel {
-}
