--- app/src/main/java/com/nexters/lettero/presentation/ui/message/viewmodel/ReadMessageViewModel.kt
@@ -1,5 +1,0 @@
-package com.nexters.lettero.presentation.ui.message.viewmodel
-
-import com.nexters.lettero.presentation.ui.base.ViewModel
-
-class ReadMessageViewModel : ViewModel {}
