--- app/src/main/java/com/nexters/lettero/presentation/ui/mypage/viewmodel/MyPageViewModel.kt
@@ -1,5 +1,0 @@
-package com.nexters.lettero.presentation.ui.mypage.viewmodel
-
-import com.nexters.lettero.presentation.ui.base.ViewModel
-
-class MyPageViewModel : ViewModel {}
