+++ app/src/main/res/layout/activity_main.xml
@@ -2,13 +2,13 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.ui.main.activity.MainActivity">
+    tools:context=".presentation.main.activity.MainActivity">
 
     <data>
 
         <variable
             name="viewModel"
-            type="com.nexters.lettero.presentation.ui.main.viewmodel.MainViewModel" />
+            type="com.nexters.lettero.presentation.main.viewmodel.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
