+++ app/src/main/res/layout/fragment_in_box.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.ui.main.fragment.InBoxFragment">
+    tools:context=".presentation.main.fragment.InBoxFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
