+++ app/src/main/res/layout/fragment_read_message.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.ui.message.fragment.ReadMessageFragment">
+    tools:context=".presentation.message.fragment.ReadMessageFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
