+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.nexters.lettero.presentation.ui.main.fragment.MainFragment"
+        android:name="com.nexters.lettero.presentation.main.fragment.MainFragment"
         android:label="fragment_main"
         tools:layout="@layout/fragment_main" >
         <action
@@ -22,7 +22,7 @@
     </fragment>
     <fragment
         android:id="@+id/introFragment"
-        android:name="com.nexters.lettero.presentation.ui.intro.fragment.IntroFragment"
+        android:name="com.nexters.lettero.presentation.intro.fragment.IntroFragment"
         android:label="fragment_intro"
         tools:layout="@layout/fragment_intro" >
         <action
@@ -34,7 +34,7 @@
     </fragment>
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.nexters.lettero.presentation.ui.login.fragment.LoginFragment"
+        android:name="com.nexters.lettero.presentation.login.fragment.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -43,17 +43,17 @@
     </fragment>
     <fragment
         android:id="@+id/myPageFragment"
-        android:name="com.nexters.lettero.presentation.ui.mypage.fragment.MyPageFragment"
+        android:name="com.nexters.lettero.presentation.mypage.fragment.MyPageFragment"
         android:label="fragment_my_page"
         tools:layout="@layout/fragment_my_page" />
     <fragment
         android:id="@+id/sendMessageFragment2"
-        android:name="com.nexters.lettero.presentation.ui.message.fragment.SendMessageFragment"
+        android:name="com.nexters.lettero.presentation.message.fragment.SendMessageFragment"
         android:label="fragment_send_message"
         tools:layout="@layout/fragment_send_message" />
     <fragment
         android:id="@+id/readMessageFragment"
-        android:name="com.nexters.lettero.presentation.ui.message.fragment.ReadMessageFragment"
+        android:name="com.nexters.lettero.presentation.message.fragment.ReadMessageFragment"
         android:label="fragment_read_message"
         tools:layout="@layout/fragment_read_message" />
 </navigation>
